Update wordbot.py
Just did a little explanatory comment edit to help people understand why there is a "header" in the request. Although all of this is redundant that is to say that it isn't necessary when getting most of the API requests. It's just that Oxford dictionary API required an User-Agent Header for me to get the data. however further down the line i got hold of a better API for this purpose and also made use of the wordhippo website for web scrapping and stuff (i don't know how "legal" all of this might be so i just ended opting for a method that doesn't require an User-Agent Header for HTTP)

"While web scraping is not inherently illegal, how it is conducted and the data's subsequent use can raise legal and ethical concerns. Actions such as scraping copyrighted content and personal information without consent or engaging in activities that disrupt the normal functioning of a website may be deemed illegal" - [From Imperva](https://www.imperva.com/blog/is-web-scraping-illegal/#:~:text=While%20web%20scraping%20is%20not,website%20may%20be%20deemed%20illegal.)